Delete blog only if id in token and blog user are same

+++ osa4/controllers/blogs.js
@@ -51,6 +51,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!(token && decodedToken.id === blog.user.toString())) {
+        return response.status(401).json({ error: 'token invalid, missing or wrong user'})
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
